App/requirements.txt
flask==3.0.3
twilio==9.0.1
python-dotenv==1.0.1
openai==1.30.1
supabase==1.0.3
gotrue==1.2.0
httpx==0.23.3
websockets==10.4
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
